// serv-admin/src/types/service.ts
export interface Service {
  id: string;
  title_uz: string;
  title_ru: string;
  descr_uz: string;
  descr_ru: string;
  btn_name_uz: string;
  btn_name_ru: string;
  btn_link: string;
  freebtn_uz: string;
  freebtn_ru: string;
  more_link: string;
  morename_uz: string;
  morename_ru: string;
  image: string;
  order: number;
  is_visible: boolean;
  created_at: string;
  updated_at: string;
}